+++ Runtime/Listeners/AnyButtonPressedListener.cs
@@ -7,13 +7,15 @@ using UnityEngine.InputSystem.Utilities;
 namespace ActionCode.InputSystem
 {
     /// <summary>
-    /// Invokes <see cref="OnAnyButtonPressed"/> event when any button is pressed once.
+    /// Invokes <see cref="OnAnyButtonPressed"/> event when any button is pressed.
     /// </summary>
     [DisallowMultipleComponent]
     public sealed class AnyButtonPressedListener : MonoBehaviour
     {
         [Min(0f), Tooltip("Time (in seconds) to wait before listen for any input.")]
         public float waitingTime = 1f;
+        [Tooltip("If enabled, it will call OnAnyButtonPressed event only one time.")]
+        public bool callOnce = true;
 
         [Space]
         [Tooltip("Event fired when any button is pressed once.")]
@@ -24,8 +26,13 @@ namespace ActionCode.InputSystem
         private void OnEnable() => Invoke(nameof(StartListenForAnyButtonPress), waitingTime);
         private void OnDisable() => anyButtonListener?.Dispose();
 
-        private void StartListenForAnyButtonPress() => anyButtonListener =
-            UnityEngine.InputSystem.InputSystem.onAnyButtonPress.CallOnce(HandleAnyButtonPressed);
+        private void StartListenForAnyButtonPress()
+        {
+            var pressEvent = UnityEngine.InputSystem.InputSystem.onAnyButtonPress;
+            anyButtonListener = callOnce ?
+                pressEvent.CallOnce(HandleAnyButtonPressed) :
+                pressEvent.Call(HandleAnyButtonPressed);
+        }
         private void HandleAnyButtonPressed(InputControl _) => OnAnyButtonPressed?.Invoke();
     }
 }